full-stack-forum-auth-service/requirements.txt
Flask
pyjwt
requests
bcrypt
python-dotenv
